fix css and ruby linter errors

+++ app/helpers/posts_helper.rb
@@ -1,25 +1,17 @@
 module PostsHelper
   def show_author(post)
-    if user_signed_in?
-      "by #{post.user.author}"
-    end
+    "by #{post.user.author}" if user_signed_in?
   end
 
   def edit_post(post)
-    if user_signed_in?
-      link_to('Edit', edit_post_path(post))
-    end
+    link_to('Edit', edit_post_path(post)) if user_signed_in?
   end
 
   def destroy_post(post)
-    if user_signed_in?
-      link_to('Destroy', post, method: :delete, data: { confirm: 'Are you sure?' })
-    end
+    link_to('Destroy', post, method: :delete, data: { confirm: 'Are you sure?' }) if user_signed_in?
   end
 
   def new_post
-    if user_signed_in?
-      link_to('New Post', new_post_path, class: "new")
-    end
+    link_to('New Post', new_post_path, class: 'new') if user_signed_in?
   end
 end
